feat: gateway함수 dm, chatroom, friend로 각각 모듈화
