+++ backend/src/app.module.ts
@@ -14,7 +14,9 @@ import {TwoFactorAuthController} from './two-factor-auth/two-factor-auth.control
 import {TwoFactorAuthService} from './two-factor-auth/two-factor-auth.service';
 import {TwoFactorAuthModule} from './two-factor-auth/two-factor-auth.module';
 import {RecordModule} from './record/record.module';
-import { GlobalVariableModule } from './global-variable/global-variable.module';
+import {GlobalVariableModule} from './global-variable/global-variable.module';
+import {DmModule} from './dm/dm.module';
+import {FriendModule} from './friend/friend.module';
 
 @Module({
   imports: [
@@ -58,6 +60,8 @@ import { GlobalVariableModule } from './global-variable/global-variable.module';
     TwoFactorAuthModule,
     RecordModule,
     GlobalVariableModule,
+    DmModule,
+    FriendModule,
   ],
   controllers: [],
   providers: [],
