+++ backend/src/chat/chat.module.ts
@@ -3,26 +3,15 @@ import {ChatGateway} from './chat.gateway';
 import {ChatController} from './chat.controller';
 import {ChatService} from './chat.service';
 import {TypeOrmModule} from '@nestjs/typeorm';
-import {
-  BlockList,
-  ChatBan,
-  ChatMember,
-  ChatRoom,
-  DirectMessage,
-  FriendList,
-} from './chat.entity';
+import {BlockList, ChatBan, ChatMember, ChatRoom} from './chat.entity';
 import {
   BlockRepository,
   ChatBanRepository,
   ChatMemberRepository,
   ChatRoomRepository,
-  DMRepository,
-  FriendListRepository,
 } from './chat.repository';
 import {User} from 'src/user/user.entitiy';
 import {UserRepository} from 'src/user/user.repository';
-import {SocketArray} from '@/global-variable/global.socket';
-import {Block} from '@/global-variable/global.block';
 import {JwtCustomModule} from '@/utils/jwt-custom.module';
 import {GlobalVariableModule} from '@/global-variable/global-variable.module';
 
@@ -33,8 +22,6 @@ import {GlobalVariableModule} from '@/global-variable/global-variable.module';
     TypeOrmModule.forFeature([ChatBan]),
     TypeOrmModule.forFeature([User]),
     TypeOrmModule.forFeature([BlockList]),
-    TypeOrmModule.forFeature([DirectMessage]),
-    TypeOrmModule.forFeature([FriendList]),
     JwtCustomModule,
     GlobalVariableModule,
   ],
@@ -46,8 +33,6 @@ import {GlobalVariableModule} from '@/global-variable/global-variable.module';
     ChatBanRepository,
     UserRepository,
     BlockRepository,
-    DMRepository,
-    FriendListRepository,
   ],
   controllers: [ChatController],
 })
